moved around methods a bit
